riscv: support kerenl to read the real-time counter by MMIO in S-mode
In the specification, the $mtime register is a memory-mapped machine-mode
register to the realtime counter provided by the platform. Because of the
permission violation, an illegal exception will occur if S-mode or U-mode
tries to read the $mtime register. The illegal exception will increase the
latency in getting the system clock and reduce the system performance.
To resolve this issue, using MMIO to read the $mtime register instead of
read CSR may be a solution because it does not issue any illegal exception
in reading $time register. To keep the M-mode authority, the kernel has to
issue an SBI call to query M-mode the appropriate method in the initialized
stage. If M-mode does not allow S-mode to read $mtime register through MMIO,
S-mode will only use csr_read(CSR_TIME) instead of MMIO.

On Sifive Unleashed board, this patch can save about 16% of latency in
handling a timer interrupt.