riscv: support using MMIO to read $mtime in vDSO
It will issue an illegal exceptions when using CSR_read($mtime) to get the
value of $mtime. This illegal exceptions will reduce the performance of
vDSO mechanism. To solve this problem, the patch make vDSO able to use
MMIO to read the $mtime register if S-mode is allowed to read $mtime
through MMIO. On unleashed board, this patch can reduce about 51% latency
in gettimeofday().