riscv: support vDSO to read $mtime via MMIO
If the platform relys on M-mode software to emulate the TIME_CSR
functionality, it will issue an illegal exceptions when using
CSR_read($mtime) to get the value of $mtime. This illegal exceptions
will reduce the performance of vDSO mechanism. To solve this problem,
this patch makes vDSO able to use MMIO to read the $mtime register if
S-mode is allowed to read $mtime through MMIO.

On unleashed board, this patch can reduce about 51% latency in
gettimeofday().

Signed-off-by: Vincent Chen <vincent.chen@sifive.com>